pyproject.toml
[project]
name = "geminiversion-proposal-writing-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.0",
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.1.3",
    "pinecone>=7.0.0",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pip>=26.0.1",
    "pytest>=8.4.1",
]
